+++ modules/statics/src/base.ts
@@ -1660,6 +1660,7 @@ export enum UnderlyingAsset {
   SRNT = 'srnt',
   SRM = 'srm',
   'sol:srm' = 'sol:srm',
+  'sol:pipe' = 'sol:pipe',
   SSV = 'ssv',
   STARS = 'stars',
   STATE = 'state',
@@ -2769,6 +2770,9 @@ export enum UnderlyingAsset {
   // BaseETH testnet tokens
   'tbaseeth:usdc' = 'tbaseeth:usdc',
 
+  // Og mainnet tokens
+  'og:wog' = 'og:wog',
+
   // Seievm mainnet tokens
   'seievm:usdc' = 'seievm:usdc',
   'seievm:weth' = 'seievm:weth',
@@ -2845,10 +2849,6 @@ export enum UnderlyingAsset {
   'erc1155:soneiumtoken' = 'erc1155:soneiumtoken',
   'tsoneium:test721' = 'tsoneium:test721',
   'tsoneium:test1155' = 'tsoneium:test1155',
-
-  // Lido ETH NFTs
-  'erc721:unsteth' = 'erc721:unsteth',
-  'terc721:unsteth' = 'terc721:unsteth',
 
   // coredao mainnet tokens
   'coredao:stcore' = 'coredao:stcore',