+++ modules/abstract-eth/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [24.13.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-eth@24.13.0...@bitgo/abstract-eth@24.13.1) (2025-09-29)
+
+**Note:** Version bump only for package @bitgo/abstract-eth
+
 # [24.13.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-eth@24.12.0...@bitgo/abstract-eth@24.13.0) (2025-09-25)
 
 ### Bug Fixes