+++ modules/sdk-coin-ada/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.15.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ada@4.15.0...@bitgo/sdk-coin-ada@4.15.1) (2025-09-29)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-ada
+
 # [4.15.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ada@4.14.0...@bitgo/sdk-coin-ada@4.15.0) (2025-09-25)
 
 ### Bug Fixes