+++ modules/sdk-coin-atom/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [13.9.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-atom@13.9.0...@bitgo/sdk-coin-atom@13.9.1) (2025-09-29)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-atom
+
 # [13.9.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-atom@13.8.2...@bitgo/sdk-coin-atom@13.9.0) (2025-09-25)
 
 ### Features