+++ modules/sdk-coin-zketh/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-zketh@2.4.0...@bitgo/sdk-coin-zketh@2.4.1) (2025-09-29)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-zketh
+
 # [2.4.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-zketh@2.3.4...@bitgo/sdk-coin-zketh@2.4.0) (2025-09-25)
 
 ### Features