+++ modules/sdk-core/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [36.10.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-core@36.9.0...@bitgo/sdk-core@36.10.0) (2025-09-29)
+
+### Bug Fixes
+
+- **sdk-core:** add keychain info to transaction verification process, fixes unit tests ([ad3c6c4](https://github.com/BitGo/BitGoJS/commit/ad3c6c47529cbb9e730d3124a33b2b2038064462))
+- **sdk-core:** fix TransactionType check ([30dbf98](https://github.com/BitGo/BitGoJS/commit/30dbf9842e908baa753b072429088167777a668f))
+- **sdk-core:** remove filter from prebuild validation ([5717147](https://github.com/BitGo/BitGoJS/commit/571714799d7c9f467ec2b216595e020f645e336b))
+- **sdk-core:** use Promise.all for concurrent processing for transaction verifying, linting fixes ([7300cc0](https://github.com/BitGo/BitGoJS/commit/7300cc04e3abf87e8b984a64eb5a1fdc47165bc1))
+
+### Features
+
+- **sdk-core:** add missing types for staking validation ([85dd4d0](https://github.com/BitGo/BitGoJS/commit/85dd4d0feac8b5f94bd5da15367bc3324d01612e))
+- **sdk-core:** add transaction verification to validate platform built transactions ([4f377a8](https://github.com/BitGo/BitGoJS/commit/4f377a8ded7d252bfaaa2bc2d6eb2ea61ff5420e))
+
 # [36.9.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-core@36.8.0...@bitgo/sdk-core@36.9.0) (2025-09-25)
 
 ### Bug Fixes