+++ modules/utxo-bin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.10.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/utxo-bin@3.10.0...@bitgo/utxo-bin@3.10.1) (2025-09-29)
+
+**Note:** Version bump only for package @bitgo/utxo-bin
+
 # [3.10.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/utxo-bin@3.9.3...@bitgo/utxo-bin@3.10.0) (2025-09-25)
 
 ### Features